Revert "Refactor memory management into separate module (#150)"
This reverts commit 56321c8b65cb66d600fd1a5aa04eace596cb9c01.
